+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Casting support for update_values of a model, similar to laravel's casts, e.g.: `uploaded_at:date`. Supports
   bool, int, float, date, string, trim (a trimmed string), none (to prevent from casting empty values to NULL)
+- Queue daemon which can run the queue reliably twice each minute, similar to the cron daemon
 
 ### 🔧 Changed